api/requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
fastapi-pagination==0.13.1
greenlet==3.2.2
h11==0.16.0
idna==3.10
pydantic==2.11.5
pydantic_core==2.33.2
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
asyncpg==0.29.0
